Ecl:Ptzp: Pattern class command update
Data getting from user on pattern API was an error.
if the data type is float or double type , this was a mistake.
Because integer converter used. This problem solved.
Now, Float converters and variables are using for pattern

+++ ptzpdriver.cpp
@@ -149,19 +149,19 @@ int PtzpDriver::set(const QString &key, const QVariant &value)
 
 	if (key == "ptz.cmd.pan_left") {
 		ptrn->commandUpdate(defaultPTHead->getPanAngle(), defaultPTHead->getTiltAngle(),
-							defaultModuleHead->getZoom(),C_PAN_LEFT, value.toInt(),0);
+							defaultModuleHead->getZoom(),C_PAN_LEFT, value.toFloat(),0);
 		defaultPTHead->panLeft(value.toFloat());
 	} else if (key == "ptz.cmd.pan_right") {
 		ptrn->commandUpdate(defaultPTHead->getPanAngle(), defaultPTHead->getTiltAngle(),
-							defaultModuleHead->getZoom(),C_PAN_RIGHT, value.toInt(),0);
+							defaultModuleHead->getZoom(),C_PAN_RIGHT, value.toFloat(),0);
 		defaultPTHead->panRight(value.toFloat());
 	} else if (key == "ptz.cmd.tilt_down") {
 		ptrn->commandUpdate(defaultPTHead->getPanAngle(), defaultPTHead->getTiltAngle(),
-							defaultModuleHead->getZoom(),C_TILT_DOWN, value.toInt(),0);
+							defaultModuleHead->getZoom(),C_TILT_DOWN, value.toFloat(),0);
 		defaultPTHead->tiltDown(value.toFloat());
 	} else if (key == "ptz.cmd.tilt_up") {
 		ptrn->commandUpdate(defaultPTHead->getPanAngle(), defaultPTHead->getTiltAngle(),
-							defaultModuleHead->getZoom(),C_TILT_UP, value.toInt(),0);
+							defaultModuleHead->getZoom(),C_TILT_UP, value.toFloat(),0);
 		defaultPTHead->tiltUp(value.toFloat());
 	} else if (key == "ptz.cmd.pan_stop") {
 		ptrn->commandUpdate(defaultPTHead->getPanAngle(), defaultPTHead->getTiltAngle(),
